portable/pyproject.toml
[project]
name = "zxf-runner"
version = "0.1.0"
description = "张雪峰转录 → 结构化知识库通用 runner，CLI 无关"
requires-python = ">=3.9"
dependencies = [
  "pyyaml>=6.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.40"]
all = ["anthropic>=0.40", "openai>=1.40"]

[project.scripts]
zxf-runner = "zxf_runner.__main__:main"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["zxf_runner*"]
